chore: rename folder from sellerProfile to market
